+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 0.1.x (October 9, 2015)
 
   * Updated README.md file,
+  * Upgraded test file to download GoeLite2 databases if they do not exist (Travis CI),
   * ...,